implement indenting
This implementation is again inspired by the JSON Indent() function
by (supposedly) rsc. But i had to rewrite it from scratch as i
estimated the structural difference of XML compared to JSON high
enough to justify this. The state machine uses a total of six states
to cope with my intuitive understanding of how XML should be
indented:

  - increase indentation if a new element starts; but only if no
    other one was closed immediately before that
  - decrease indentation if an element ends; but only if one ended
    before that

There are a lot of state transitions going on to achieve this and i
have the slight suspicion it might be implemented more concise...